[ios] Introduce visible annotations API
This adds an API for getting all annotations that are currently
visible on the map. In the future, an additional API can be added
to query by a specific rect. That API can be used internally to
power this one.

This also adds a new option to iosapp to test the feature. At the
time of this commit, the underlying `queryPointAnnotations` does
not appear to work as expected.